Crate pythagorean_triplet

The solution for Exercism “Pythagorean triplet” problem.

Requires the nightly compiler where Iterator::flatten is stabilized. Tested on rustc 1.29.0-nightly (9fd3d7899 2018-07-07).

Description

A Pythagorean triplet is a set of three natural numbers, (a, b, c), for which,

a**2 + b**2 = c**2

For example,

3**2 + 4**2 = 9 + 16 = 25 = 5**2.

There exists exactly one Pythagorean triplet for which a + b + c = 1000.

Find the product a * b * c.

Terms used throughput the crate

Pythagorean triple:
Three natural numbers, a, b, and c, such that a2 + b2 = c2. A well-known example is (3, 4, 5).

Primitive Pythagorean triple:
A Pythagorean triple (a, b, c), such that gcd(a, b, c) = 1.

Structs

PrimitiveTriples
An infinite iterator over primitive Pythagorean triples.

Triple
Represents a Pythagorean triple.

Triples
An infinite iterator over triples derived from a primitive Pythagorean triple.

Functions

find
Tries to find a Pythagorean triple (a, b, c), such that a + b + c = 1000.